tiny_segment_app/requirements.txt
django
django-bootstrap5
django-crispy-forms
whitenoise
psycopg2
psycopg2-binary
django-htmx
gunicorn
celery
redis
pika
faker
requests
openai
djangorestframework